Refactor fetch functions, change name for DELETE function
