+++ src/components/EventList.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable  @typescript-eslint/no-non-null-assertion */
 import NextLink from 'next/link';
 import {
     Button,
@@ -11,8 +12,8 @@ import {
     FormattedMessage as Msg,
 } from 'react-intl';
 
+import deleteEventResponse from '../fetching/deleteEventResponse';
 import putEventResponse from '../fetching/putEventResponse';
-import putUndoEventResponse from '../fetching/putUndoEventResponse';
 import { ZetkinEvent } from '../interfaces/ZetkinEvent';
 import { ZetkinOrganization } from '../interfaces/ZetkinOrganization';
 
@@ -39,7 +40,7 @@ const EventList = ({ events, org } : EventListProps) : JSX.Element => {
                         <View data-test="event-title">
                             { e.title ? e.title : e.activity.title }
                         </View>
-                        <View data-test="org-title">{ org?.title }</View>
+                        <View data-test="org-title">{ org!.title }</View>
                         <View data-test="campaign-title">{ e.campaign.title }</View>
                         <View data-test="start-time">
                             <FormattedDate
@@ -65,18 +66,18 @@ const EventList = ({ events, org } : EventListProps) : JSX.Element => {
                         <Button
                             data-test="sign-up-button"
                             marginTop="size-50"
-                            onPress={ () => putEventResponse(e.id, org?.id) }
+                            onPress={ () => putEventResponse(e.id, org!.id) }
                             variant="cta">
                             <Msg id="misc.eventList.signup"/>
                         </Button>
                         <Button
                             data-test="undo-sign-up-button"
                             marginTop="size-50"
-                            onPress={ () => putUndoEventResponse(e.id, org?.id) }
+                            onPress={ () => deleteEventResponse(e.id, org!.id) }
                             variant="cta">
                             <Msg id="misc.eventList.undoSignup"/>
                         </Button>
-                        <NextLink href={ `/o/${org?.id}/events/${e.id}` }>
+                        <NextLink href={ `/o/${org!.id}/events/${e.id}` }>
                             <a>
                                 <Button marginTop="size-50" variant="cta">
                                     <Msg id="misc.eventList.moreInfo"/>
