+++ src/types/zetkin.ts
@@ -7,3 +7,13 @@ export interface ZetkinMembership {
         id: number;
     };
 }
+
+export interface ZetkinEventResponse {
+    action_id: number;
+    respone_date: string;
+    person: {
+        name: string;
+        id: number;
+    };
+    id: number;
+}
